player.remoteTextTracks -> player.textTracks
Plugin was throwing an error looking for player.remoteTextTracks method. I was unable to locate the method in the videojs library, but upon searching through old commits on this file it looks like the method is a remnant that wasn't corrected.

If I am wrong, I'd like to know where to find the method remoteTextTracks on the player object.