Update current content video to be sourced from `snapshot.source` instead of `snapshot.src` and `snapshot.type`.
Remove `snapshot.type` since it's unnecessary. I also suppose there was an existing issue where the `type` did not exist, but I am guessing that it was a silent error.
This will no longer be a problem as the current content video will be initialized with the exact same data that it was original configured with.

Minor formatting.